settings files refactoring
