+++ config/config.json
@@ -1,1 +1,1 @@
-{"master_height": "500", "master_width": "800", "prev_window_size": 100, "preview_size": 500, "txt_mastertitle": "Search Engine", "txt_selectdir": "Select Directory", "txt_settingsheader": "Settings", "txt_entrysearch": "Search", "txt_entryclear": "Clear", "txt_resultitems": "Search result", "txt_preview": "Preview", "txt_page": "Page", "txt_okay": "Ok", "txt_exitpreview": "Exit Preview", "txt_language": "Language: ", "txt_colortheme": "Theme: ", "txt_selectStemmer": "Select Stemmer", "txt_toggleStopword": "Stop Word Removal", "txt_on": "on", "txt_off": "off", "ERR_noDirectorySelected": "ERROR: No directory selected", "ERR_resultListEmpty": "ERROR: There are no search results", "col_bg": "#3b3b3b", "col_bg_lgt": "#5f5f5f", "col_btn_idle": "#940000", "col_btn_active": "#d50000", "col_entryfield_idle": "#bbbbbb", "col_entryfield_contrast": "#3b3b3b", "col_acc_major": "#b3b3b3", "col_acc_minor": "#b3b3b3", "font_header_1": ["Arial", 15, "bold"], "font_header_2": ["Arial", 10, "bold"], "font_returntext": ["Arial", 10], "relief_frames": "flat", "relief_btn": "flat", "stemmer": "porter", "stop_word": true}
+{"master_height": "500", "master_width": "800", "prev_window_size": 100, "preview_size": 500, "txt_mastertitle": "Suchmaschine", "txt_selectdir": "Verzeichnis ausw\u00e4hlen", "txt_settingsheader": "Einstellungen", "txt_entrysearch": "Suchen", "txt_entryclear": "L\u00f6schen", "txt_resultitems": "Ergebnisse", "txt_preview": "Vorschau", "txt_page": "Seite", "txt_okay": "Ok", "txt_exitpreview": "Schlie\u00dfen", "txt_language": "Sprache: ", "txt_colortheme": "Design: ", "txt_selectStemmer": "Stemmerauswahl", "txt_toggleStopword": "Stoppwortentfernung", "txt_on": "an", "txt_off": "aus", "ERR_noDirectorySelected": "FEHLER: Kein Verzeichnis ausgew\u00e4hlt", "ERR_resultListEmpty": "FEHLER: Keine Suchergebnisse vorhanden", "col_bg": "#3b3b3b", "col_bg_lgt": "#5f5f5f", "col_btn_idle": "#940000", "col_btn_active": "#d50000", "col_entryfield_idle": "#bbbbbb", "col_entryfield_contrast": "#3b3b3b", "col_acc_major": "#b3b3b3", "col_acc_minor": "#b3b3b3", "font_header_1": ["Arial", 15, "bold"], "font_header_2": ["Arial", 10, "bold"], "font_returntext": ["Arial", 10], "relief_frames": "flat", "relief_btn": "flat", "stemmer": "lancaster", "stop_word": false}
