+++ gui/builder_toolbox/settings_defaultpaths.py
@@ -1,0 +1,4 @@
+default_path = "./config"
+masterconfigfile = "/config.json"
+languageconfigfile = "/config_lang.json"
+colorsconfigfile = "/config_colors.json"
